Move project creation logic to util
